+++ stack/app/variables.tf
@@ -139,6 +139,12 @@ variable "skip_elasticache" {
   default = false
 }
 
+variable "elasticache_transit_encryption_enabled" {
+  type        = bool
+  default     = true
+  description = ":warning: changing this from `false` to `true` requires a re-creation of the cluster"
+}
+
 variable "elasticache_name" {
   type    = string
   default = null
@@ -300,6 +306,15 @@ variable "health_check_path" { default = "/livez" }
 variable "allow_internal_traffic_to_ports" {
   type    = list(string)
   default = []
+}
+
+variable "keep_alive_timeout" {
+  type    = number
+  default = 60
+  validation {
+    condition     = var.keep_alive_timeout >= 60 && var.keep_alive_timeout <= 4000
+    error_message = "keep_alive_timeout must be between 60 and 4000"
+  }
 }
 
 variable "allow_alb_traffic_to_ports" {