+++ CHANGELOG.md
@@ -18,6 +18,9 @@ This project does not follow SemVer, since modules are independent of each other
 - make keep alive timeout configurable, and fix default for NLB as well (previously only fixed for ALB). [#232](https://github.com/dbl-works/terraform/pull/232)
 - pass through the keep alive timeout to the `stack` module. [#233](https://github.com/dbl-works/terraform/pull/232)
 
+### Stack
+- pass through the keep alive timeout to the `stack` module. [#234](https://github.com/dbl-works/terraform/pull/234)
+
 ## [v2023.07.19]
 ### ecs-deploy
 - Allow ecs service to be deployed to private subnets. [#228](https://github.com/dbl-works/terraform/pull/228)