+++ CHANGELOG.md
@@ -8,6 +8,21 @@ This project does not follow SemVer, since modules are independent of each other
 
 ## Unreleased
 ### ecs-deploy
+- Add deployment_circuit_breaker options [#230](https://github.com/dbl-works/terraform/pull/230)
+
+### elasticache
+- Enable `encryption-in-transit` by default, :warning: requires re-creation of the cluster (you can thus opt-out) [#231](https://github.com/dbl-works/terraform/pull/231)
+- Change default Redis version from `6` to `7` [#231](https://github.com/dbl-works/terraform/pull/231)
+
+### ECS
+- make keep alive timeout configurable, and fix default for NLB as well (previously only fixed for ALB). [#232](https://github.com/dbl-works/terraform/pull/232)
+- pass through the keep alive timeout to the `stack` module. [#233](https://github.com/dbl-works/terraform/pull/232)
+
+### Stack
+- pass through the keep alive timeout to the `stack` module. [#234](https://github.com/dbl-works/terraform/pull/234)
+
+## [v2023.07.19]
+### ecs-deploy
 - Allow ecs service to be deployed to private subnets. [#228](https://github.com/dbl-works/terraform/pull/228)
 
 ### fivetran/connectors/lambda
@@ -35,6 +50,7 @@ This project does not follow SemVer, since modules are independent of each other
 
 ### ecs
 - add alb_listener_rule and allow_alb_traffic_to_ports options, remove subnet_private_ids since it is not used [#216](https://github.com/dbl-works/terraform/pull/216)
+- Set Load Balancer keep alive timeout to 60 seconds [#229](https://github.com/dbl-works/terraform/pull/229)
 
 ### stack/app
 - add alb_listener_rule and allow_alb_traffic_to_ports options [#216](https://github.com/dbl-works/terraform/pull/216)